add gigglebot module & documentation & optimize RAM usage [ci skip]
