+++ .gitignore
@@ -1,0 +1,4 @@
+build/
+docs/build
+.vscode/
+*.pyc
