+++ docs/source/addon-sensors-tutorial.rst
@@ -1,0 +1,3 @@
+#######################
+Add-On Sensors Tutorial
+#######################
