+++ docs/source/code.rst
@@ -1,6 +1,17 @@
+####
 API
-============================
+####
+
+*********
+GiggleBot
+*********
 
 .. automodule:: gigglebot
    :members:
 
+***************
+Distance Sensor
+***************
+
+.. autoclass:: distance_sensor.DistanceSensor
+   :members:
